note taker working
