+++ routes/html_routes.js
@@ -1,0 +1,12 @@
+const router = require('express').Router();
+const path = require('path');
+
+router.use('/notes', (req, res) => {
+    res.sendFile(path.join(__dirname, '../public/notes.html'))
+});
+ 
+router.use('/', (req, res) => {
+    res.sendFile(path.join(__dirname, '../public/index.html'))
+});
+
+module.exports = router;
